Highlight existing code for work order   - add router   - when land to caut contact page from w-o page show highlights of existing codes   - when land to caut contact page from index page, no hightlights
