+++ src/styles/all.scss
@@ -32,3 +32,4 @@ $govuk-font-family: 'GDS Transport', arial, sans-serif !important;
 @import 'components/operative/work-orders';
 @import 'components/grid';
 @import 'components/info-modal';
+@import 'components/caut-contact/caut-contact.scss';
