+++ src/styles/components/caut-contact/caut-contact.scss
@@ -1,0 +1,4 @@
+.highlight-for-existing-code {
+  color: repairs-hub-colour('dark-red');
+  font-weight: bold;
+}
